Lab1 Fix2

+++ LibraryCPP/Tests/MyArray.cpp
@@ -15,7 +15,7 @@ int main(int argc, char **argv)
 
     bool isIncorrect = false;
 
-    for(int i = 0; i < array_size(arr); i++)
+    for(size_t i = 0; i < array_size(arr); i++)
     {
         if(array_get(arr,i) != array_get(correctAns,i))
         {
